innovation-lab-examples/gemini-quickstart/07-short-film-agent/unified-movie-agent/requirements.txt
uagents>=0.20.0
uagents-core>=0.15.0
google-genai>=1.0.0
google-cloud-storage>=2.14.0
google-auth>=2.25.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
stripe>=8.0.0
